+++ R/write_output.R
@@ -46,7 +46,19 @@ group_sistec <- function(x){
   sistec_without_rfept <- x$sistec_without_rfept %>%
     dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "sistec_without_rfept")
   
-  dplyr::full_join(sistec_without_cpf, sistec_without_rfept, by = c("CAMPUS", "CURSO"))
+  sistec_wrong_cpf <- x$sistec_wrong_cpf %>%
+    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "sistec_wrong_cpf")
+  
+  sistec_duplicated_registry <- x$sistec_duplicated_registry %>%
+    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "sistec_duplicated_registry")
+  
+  sistec_pending <- x$sistec_pending %>%
+    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "sistec_pending")
+  
+  dplyr::full_join(sistec_without_cpf, sistec_without_rfept, by = c("CAMPUS", "CURSO")) %>%
+    dplyr::full_join(sistec_wrong_cpf, by = c("CAMPUS", "CURSO")) %>% 
+    dplyr::full_join(sistec_duplicated_registry, by = c("CAMPUS", "CURSO")) %>% 
+    dplyr::full_join(sistec_pending, by = c("CAMPUS", "CURSO"))
 }
 
 #' @importFrom dplyr %>% sym
@@ -57,12 +69,21 @@ group_rfept <- function(x){
   
   rfept_without_sistec <- x$rfept_without_sistec %>%
     dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "rfept_without_sistec")
-
-  rfept_wrong_beginning <- x$rfept_wrong_beginning %>%
-    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "rfept_wrong_beginning")
   
-  rfept_wrong_cyclo <- x$rfept_wrong_cyclo %>%
-    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "rfept_wrong_cyclo")
+  rfept_wrong_cpf <- x$rfept_wrong_cpf %>%
+    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "rfept_wrong_cpf")
+  
+  rfept_duplicated_registry <- x$rfept_duplicated_registry %>%
+    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "rfept_duplicated_registry")
+  
+  rfept_pending <- x$rfept_pending %>%
+    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "rfept_pending")
+  
+  # rfept_wrong_cyclo <- x$rfept_wrong_cyclo %>%
+  #   dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "rfept_wrong_cyclo")
+  
+  wrong_beginning <- x$wrong_beginning %>%
+    dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "wrong_beginning")
   
   situation_to_update <- x$situation_to_update %>%
     dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "situation_to_update")
@@ -71,8 +92,10 @@ group_rfept <- function(x){
     dplyr::group_nest(!!sym("CAMPUS"), !!sym("CURSO"), .key = "situation_updated")  
  
   dplyr::full_join(rfept_without_cpf, rfept_without_sistec, by = c("CAMPUS", "CURSO")) %>%
-    dplyr::full_join(rfept_wrong_beginning, by = c("CAMPUS", "CURSO")) %>% 
-    dplyr::full_join(rfept_wrong_cyclo, by = c("CAMPUS", "CURSO")) %>% 
+    dplyr::full_join(rfept_wrong_cpf, by = c("CAMPUS", "CURSO")) %>% 
+    dplyr::full_join(rfept_duplicated_registry, by = c("CAMPUS", "CURSO")) %>% 
+    dplyr::full_join(rfept_pending, by = c("CAMPUS", "CURSO")) %>% 
+    dplyr::full_join(wrong_beginning, by = c("CAMPUS", "CURSO")) %>% 
     dplyr::full_join(situation_to_update, by = c("CAMPUS", "CURSO")) %>% 
     dplyr::full_join(situation_updated, by = c("CAMPUS", "CURSO"))
 }
@@ -90,8 +113,11 @@ write_sistec <- function(x, path, rfept_table){
       path_to_save <- paste0(path, "/Retificar no Sistec/",  x$CAMPUS[e], "/", x$CURSO[e])
       dir.create(path_to_save, recursive = TRUE)
       sistec <- list()
-      sistec[["Retificar CPF"]] <- x$sistec_without_cpf[e][[1]]
+      sistec[["Sem CPF"]] <- x$sistec_without_cpf[e][[1]]
+      sistec[["Erro no CPF"]] <- x$sistec_wrong_cpf[e][[1]]
+      sistec[["Entrada duplicada"]] <- x$sistec_duplicated_registry[e][[1]]
       sistec[[paste0("Inserir no ", rfept_table)]] <- x$sistec_without_rfept[e][[1]]
+      sistec[["Inspe\u00e7\u00e3o manual"]] <- x$sistec_pending[e][[1]]
       openxlsx::write.xlsx(sistec, paste0(path_to_save, "/Sistec.xlsx"))
     })
   )
@@ -103,10 +129,12 @@ write_rfept <- function(x, path, rfept_table){
       path_to_save <- paste0(path, "/Retificar no ", rfept_table, "/",  x$CAMPUS[e], "/", x$CURSO[e])
       dir.create(path_to_save, recursive = TRUE)
       rfept <- list()
-      rfept[["Retificar CPF"]] <- x$rfept_without_cpf[e][[1]]
+      rfept[["Sem CPF"]] <- x$rfept_without_cpf[e][[1]]
+      rfept[["Erro no CPF"]] <- x$rfept_wrong_cpf[e][[1]]
+      rfept[["Entrada duplicada"]] <- x$rfept_duplicated_registry[e][[1]]
       rfept[["Inserir no Sistec"]] <- x$rfept_without_sistec[e][[1]]
-      rfept[["Retificar Inicio"]] <- x$rfept_wrong_beginning[e][[1]]
-      rfept[["Retificar Ciclo"]] <- x$rfept_wrong_cyclo[e][[1]]
+      rfept[["Inspe\u00e7\u00e3o manual"]] <- x$rfept_pending[e][[1]]
+      rfept[["Retificar Inicio"]] <- x$wrong_beginning[e][[1]]
       rfept[["Retificar Situa\u00e7\u00e3o"]] <- x$situation_to_update[e][[1]]
       rfept[["Situa\u00e7\u00f5es Atualizadas"]] <- x$situation_updated[e][[1]]
       openxlsx::write.xlsx(rfept, paste0(path_to_save, "/",