+++ NEWS.md
@@ -3,6 +3,11 @@
 ### Functions
 
 - Update `read_sigaa()`to read datasets with "Campus" and "Cota". (#53)
+- Cration of `read_linked_courses()` to read a file with the relation between courses in academic registration and Sistec cyclo. (#55)
+
+### ARIA
+
+- New to button to include linked courses files. (#55)
 
 # sistec 0.1.0